// src/orval/type.schemas.ts
export type RequestType = typeof RequestType[keyof typeof RequestType];



export const RequestType = {
  Default: 'Default',
  Buy: 'Buy',
  Sell: 'Sell',
  Rent: 'Rent',
  RentOut: 'RentOut',
  Insurance: 'Insurance',
  Dealer: 'Dealer',
  Collaborator: 'Collaborator',
  Advertisement: 'Advertisement',
  Enquiry: 'Enquiry',
} as const;

export type RequestStatus = typeof RequestStatus[keyof typeof RequestStatus];



export const RequestStatus = {
  Default: 'Default',
  Pending: 'Pending',
  Appointment: 'Appointment',
  PendingPayment: 'PendingPayment',
  PendingDelivery: 'PendingDelivery',
  Completed: 'Completed',
  Incomplete: 'Incomplete',
} as const;
// ...
export type ListingType = typeof ListingType[keyof typeof ListingType];



export const ListingType = {
  Default: 'Default',
  Sell: 'Sell',
  Rent: 'Rent',
} as const;
// ...
export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];



export const AppointmentStatus = {
  Default: 'Default',
  Scheduled: 'Scheduled',
  Completed: 'Completed',
  Cancelled: 'Cancelled',
} as const;